Deduplicate Amp/Pi adapters and hide Pi from personal credentials
Extract the near-identical Amp/Pi streaming-JSON parser and Execute
flow into stream_parser.go driven by a streamParseConfig so each
adapter only declares its per-agent config. Fix two parsing bugs
uncovered by the refactor: result events that report only output
tokens are now recorded (input=0 is legitimate on continuation
turns), and error events with no error/message/content fields emit
"unknown error" instead of a blank log line. Add an AMP_MODE shell
injection test mirroring Pi's. On the frontend, tag Pi with
inheritsProviderKeys so the personal-credential selector skips
meta-agents that have no key to save, and drop the dead pi entry
from KEY_PLACEHOLDERS.
