+++ internal/services/agent/adapters/stream_parser.go
@@ -1,0 +1,300 @@
+package adapters
+
+import (
+	"bytes"
+	"context"
+	"encoding/json"
+	"fmt"
+	"strings"
+	"time"
+
+	"github.com/rs/zerolog"
+
+	"github.com/assembledhq/143/internal/services/agent"
+)
+
+// agentStreamEvent is the union of fields emitted by the amp and pi stream-JSON
+// protocols (both Claude Code-compatible). Fields not present on a given event
+// are zero-valued.
+type agentStreamEvent struct {
+	Type      string          `json:"type"`
+	Content   string          `json:"content,omitempty"`
+	Message   string          `json:"message,omitempty"`
+	Tool      string          `json:"tool,omitempty"`
+	Name      string          `json:"name,omitempty"`
+	Input     json.RawMessage `json:"input,omitempty"`
+	Output    string          `json:"output,omitempty"`
+	Result    json.RawMessage `json:"result,omitempty"`
+	Error     string          `json:"error,omitempty"`
+	Model     string          `json:"model,omitempty"`
+	SessionID string          `json:"session_id,omitempty"`
+	Usage     *struct {
+		InputTokens  int `json:"input_tokens"`
+		OutputTokens int `json:"output_tokens"`
+	} `json:"usage,omitempty"`
+}
+
+// streamParseConfig captures the small per-adapter differences in event
+// dispatch. Everything else is identical between amp and pi.
+type streamParseConfig struct {
+	// MessageAsAssistant routes a type=="message" event through the assistant
+	// branch (content → output log, updates lastAssistant).
+	MessageAsAssistant bool
+	// DoneAsResult routes a type=="done" event through the result/usage branch.
+	DoneAsResult bool
+	// CaptureToolModel records event.Model in tool_use metadata.
+	CaptureToolModel bool
+	// CaptureSessionID records event.SessionID on the shared AgentResult when
+	// a result/usage event carries one.
+	CaptureSessionID bool
+}
+
+// parseAgentStreamLine processes a single line of streaming JSON output for an
+// agent that emits Claude Code-compatible events. Unknown event types fall
+// through to a debug log; non-JSON lines are emitted as plain output and
+// scanned for confidence markers.
+func parseAgentStreamLine(
+	line []byte,
+	cfg streamParseConfig,
+	result *agent.AgentResult,
+	logCh chan<- agent.LogEntry,
+	summaryParts *[]string,
+	lastAssistant *string,
+) {
+	var event agentStreamEvent
+	if err := json.Unmarshal(line, &event); err != nil {
+		text := string(line)
+		logCh <- agent.LogEntry{
+			Timestamp: time.Now(),
+			Level:     "output",
+			Message:   text,
+		}
+		*summaryParts = append(*summaryParts, text)
+		tryExtractConfidence(text, result)
+		return
+	}
+
+	switch {
+	case event.Type == "assistant" || event.Type == "text" ||
+		(cfg.MessageAsAssistant && event.Type == "message"):
+		content := event.Content
+		if content == "" {
+			content = event.Message
+		}
+		logCh <- agent.LogEntry{
+			Timestamp: time.Now(),
+			Level:     "output",
+			Message:   content,
+		}
+		*lastAssistant = content
+		tryExtractConfidence(content, result)
+
+	case event.Type == "tool_use" || event.Type == "tool_call":
+		toolName := event.Tool
+		if toolName == "" {
+			toolName = event.Name
+		}
+		metadata := map[string]interface{}{"tool": toolName}
+		if cfg.CaptureToolModel && event.Model != "" {
+			metadata["model"] = event.Model
+		}
+		if len(event.Input) > 0 {
+			var inputMap map[string]interface{}
+			if err := json.Unmarshal(event.Input, &inputMap); err == nil {
+				metadata["input"] = inputMap
+			}
+		}
+		logCh <- agent.LogEntry{
+			Timestamp: time.Now(),
+			Level:     "tool_use",
+			Message:   fmt.Sprintf("using tool: %s", toolName),
+			Metadata:  metadata,
+		}
+
+	case event.Type == "tool_result":
+		toolName := event.Tool
+		if toolName == "" {
+			toolName = event.Name
+		}
+		metadata := map[string]interface{}{"type": "tool_result"}
+		if toolName != "" {
+			metadata["tool"] = toolName
+		}
+		outputMsg := event.Output
+		if outputMsg == "" && len(event.Result) > 0 {
+			outputMsg = string(event.Result)
+		}
+		logCh <- agent.LogEntry{
+			Timestamp: time.Now(),
+			Level:     "output",
+			Message:   outputMsg,
+			Metadata:  metadata,
+		}
+
+	case event.Type == "thinking":
+		logCh <- agent.LogEntry{
+			Timestamp: time.Now(),
+			Level:     "debug",
+			Message:   event.Content,
+			Metadata:  map[string]interface{}{"type": "thinking"},
+		}
+
+	case event.Type == "error":
+		msg := event.Error
+		if msg == "" {
+			msg = event.Message
+		}
+		if msg == "" {
+			msg = event.Content
+		}
+		if msg == "" {
+			msg = "unknown error"
+		}
+		logCh <- agent.LogEntry{
+			Timestamp: time.Now(),
+			Level:     "error",
+			Message:   msg,
+		}
+
+	case event.Type == "result" || event.Type == "usage" ||
+		(cfg.DoneAsResult && event.Type == "done"):
+		content := event.Content
+		logCh <- agent.LogEntry{
+			Timestamp: time.Now(),
+			Level:     "info",
+			Message:   content,
+		}
+		if content != "" {
+			*summaryParts = append(*summaryParts, content)
+			tryExtractConfidence(content, result)
+		}
+		// Both shapes ship in the wild: a dedicated `usage` object, and a
+		// `result` payload that sometimes packs the same counters. Accept
+		// either; `result` is checked last so it wins when both are present.
+		if event.Usage != nil {
+			result.TokenUsage = agent.TokenUsage{
+				InputTokens:  event.Usage.InputTokens,
+				OutputTokens: event.Usage.OutputTokens,
+			}
+		}
+		if len(event.Result) > 0 {
+			var usage agent.TokenUsage
+			if err := json.Unmarshal(event.Result, &usage); err == nil &&
+				(usage.InputTokens > 0 || usage.OutputTokens > 0) {
+				result.TokenUsage = usage
+			}
+		}
+		if cfg.CaptureSessionID && event.SessionID != "" {
+			result.AgentSessionID = event.SessionID
+		}
+
+	default:
+		logCh <- agent.LogEntry{
+			Timestamp: time.Now(),
+			Level:     "debug",
+			Message:   string(line),
+		}
+	}
+}
+
+// streamingAgentConfig drives runStreamingAgent. BuildCmd receives the already
+// shell-escaped prompt file path and returns the full command line.
+type streamingAgentConfig struct {
+	DisplayName string // "Amp" / "Pi" — used in start/completion log messages.
+	CLIName     string // "amp" / "pi" — used in the exec error message.
+	BuildCmd    func(escapedPromptPath string) string
+	ParseConfig streamParseConfig
+}
+
+// runStreamingAgent implements the shared Execute flow for agents that (a)
+// take their prompt via a file on disk, (b) emit Claude Code-compatible
+// stream JSON, and (c) don't have a headless continuation flag.
+func runStreamingAgent(
+	ctx context.Context,
+	cfg streamingAgentConfig,
+	logger zerolog.Logger,
+	sandbox *agent.Sandbox,
+	prompt *agent.AgentPrompt,
+	logCh chan<- agent.LogEntry,
+) (*agent.AgentResult, error) {
+	provider := agent.SandboxProviderFromContext(ctx)
+	if provider == nil {
+		return nil, fmt.Errorf("sandbox provider not found in context")
+	}
+
+	var promptContent string
+	if prompt.Continuation {
+		promptContent = prompt.UserMessage
+	} else {
+		promptContent = fmt.Sprintf("%s\n\n---\n\n%s", prompt.SystemPrompt, prompt.UserPrompt)
+	}
+	promptPath := fmt.Sprintf("%s/.143-prompt.md", sandbox.WorkDir)
+	if err := provider.WriteFile(ctx, sandbox, promptPath, []byte(promptContent)); err != nil {
+		return nil, fmt.Errorf("write prompt file: %w", err)
+	}
+
+	cmd := cfg.BuildCmd(shellEscapeSingle(promptPath))
+
+	logCh <- agent.LogEntry{
+		Timestamp: time.Now(),
+		Level:     "info",
+		Message:   fmt.Sprintf("starting %s CLI", cfg.DisplayName),
+		Metadata:  map[string]interface{}{"max_tokens": prompt.MaxTokens, "resume": prompt.Continuation},
+	}
+
+	result := &agent.AgentResult{}
+	var stderr bytes.Buffer
+	var summaryParts []string
+	var lastAssistantContent string
+
+	exitCode, err := provider.ExecStream(ctx, sandbox, cmd, func(line []byte) {
+		if len(bytes.TrimSpace(line)) == 0 {
+			return
+		}
+		parseAgentStreamLine(line, cfg.ParseConfig, result, logCh, &summaryParts, &lastAssistantContent)
+	}, &stderr)
+	if err != nil {
+		return nil, fmt.Errorf("exec %s CLI: %w", cfg.CLIName, err)
+	}
+
+	result.ExitCode = exitCode
+	if len(summaryParts) > 0 {
+		result.Summary = strings.Join(summaryParts, "\n")
+	} else if lastAssistantContent != "" {
+		result.Summary = lastAssistantContent
+	}
+
+	if stderr.Len() > 0 {
+		logCh <- agent.LogEntry{
+			Timestamp: time.Now(),
+			Level:     "error",
+			Message:   stderr.String(),
+		}
+	}
+
+	if exitCode != 0 {
+		result.Error = fmt.Sprintf("%s CLI exited with code %d", cfg.CLIName, exitCode)
+		if stderr.Len() > 0 {
+			result.Error += ": " + stderr.String()
+		}
+	}
+
+	diff, err := collectDiff(ctx, provider, sandbox)
+	if err != nil {
+		logger.Warn().Err(err).Msg("failed to collect git diff")
+	} else {
+		result.Diff = diff
+	}
+
+	logCh <- agent.LogEntry{
+		Timestamp: time.Now(),
+		Level:     "info",
+		Message:   fmt.Sprintf("%s CLI completed", cfg.DisplayName),
+		Metadata: map[string]interface{}{
+			"exit_code":        exitCode,
+			"confidence_score": result.ConfidenceScore,
+		},
+	}
+
+	return result, nil
+}
